Extract context into directories and update file paths
