+++ client/src/components/Alert.js
@@ -1,6 +1,6 @@
 import React, {useContext} from 'react'
 import "../styles/alerts.css"
-import AuthContext from "../context/AuthContext"
+import AuthContext from "../context/authentication/AuthContext"
 import ErrorIcon from '@material-ui/icons/Error';
 import CheckCircleIcon from '@material-ui/icons/CheckCircle';
 import {CSSTransition, TransitionGroup} from "react-transition-group"
