+++ client/src/components/Content.js
@@ -1,7 +1,7 @@
 import React, {useContext, useEffect} from 'react'
 import Input from './Input'
 import List from './List';
-import AppContext from "../context/AppContext";
+import AppContext from "../context/application/AppContext";
 import { CircularProgress } from '@material-ui/core';
 
 const Content = () => {
