+++ client/src/components/Datepicker.js
@@ -1,5 +1,5 @@
 import React, { useContext, useState } from 'react';
-import AppContext from "../context/AppContext";
+import AppContext from "../context/application/AppContext";
 import CalendarTodayIcon from '@material-ui/icons/CalendarToday';
 import DatePicker from "react-datepicker"
 import "react-datepicker/dist/react-datepicker.css";
