+++ client/src/components/EditProfile.js
@@ -3,7 +3,7 @@ import Alert from "../components/Alert"
 import { Link } from 'react-router-dom'
 import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos';
 import PersonIcon from '@material-ui/icons/Person';
-import AuthContext from "../context/AuthContext"
+import AuthContext from "../context/authentication/AuthContext"
 
 import pageFlip from "../sounds/page-flip.wav"
 import EditInfo from './EditInfo';
