+++ client/src/components/RegisterModal.js
@@ -1,5 +1,5 @@
 import React, {useContext, useState} from 'react';
-import AuthContext from "../context/AuthContext";
+import AuthContext from "../context/authentication/AuthContext";
 
 const RegisterModal = (props) => {
     const authContext = useContext(AuthContext);
