+++ client/src/context/authentication/AuthReducer.js
@@ -6,7 +6,7 @@ import {
     USER_LOADED, 
     SET_ALERT,
     REMOVE_ALERT 
-} from "./types";
+} from "../types/types";
 
 //eslint-disable-next-line
 export default (state, action) => {
